Refactor endpoints into Blueprints
