+++ feedback/backend/routes/__init__.py
@@ -1,0 +1,7 @@
+from .admin import admin_bp
+from .feedback import feedback_bp
+from .user import user_bp
+
+admin_bp = admin_bp
+feedback_bp = feedback_bp
+user_bp = user_bp
